# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Udemy API'
  description: 'BY Phan Văn Tuấn '
  contact:
    name: 'Swagger API Team'
  version: '1.0'
servers:
  -
    url: 'http://localhost:8000'
    description: 'API server'
  -
    url: 'https://api-study.salyr.online'
    description: 'production server'
paths:
...
  '/api/categories/{categoryId}/courses':
    get:
      tags:
        - Categories
      summary: 'Get Courses by Category'
      description: 'Returns all courses under a specific category'
      operationId: getCourses
      parameters:
        -
          name: categoryId
          in: path
          description: 'ID of category to return courses for'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseResource'
        404:
          description: 'Category not found'
...
  /api/lectures:
    get:
      tags:
        - lectures
      summary: 'List all lectures'
      description: 'Display a listing of the lectures'
      operationId: 9e761e5869f01f3dab9a5ea6cd5cf0fa
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LectureResource'
    post:
      tags:
        - lectures
      summary: 'Create a lecture'
      description: 'Store a newly created lecture in storage'
      operationId: 885111951104828aad3a2a12f19a93ad
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StoreLectureRequest'
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LectureResource'
        500:
          description: 'Error creating lecture'
      security:
        -
          bearerAuth: []
  '/api/lectures/{id}':
    get:
      tags:
        - lectures
      summary: 'Show a specific lecture'
      description: 'Display the specified lecture'
      operationId: cba6066aacd9cd43738aff77eab50c4b
      parameters:
        -
          name: id
          in: path
          description: 'ID of lecture to return'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Successful operation'
        404:
          description: 'Lecture not found'
    post:
      tags:
        - lectures
      summary: 'Update a lecture'
      description: 'Update the specified lecture in storage'
      operationId: dbd9063b23cddf7620b91338f22066af
      parameters:
        -
          name: id
          in: path
          description: 'ID of lecture to delete'
          required: true
          schema:
            type: integer
        -
          name: _method
          in: query
          description: 'Method to be used'
          required: false
          schema:
            type: string
            default: PUT
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateLectureRequest'
      responses:
        200:
          description: 'Lecture updated successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LectureResource'
        404:
          description: 'Lecture not found'
        422:
          description: 'Invalid input'
        500:
          description: 'Error updating lecture'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - lectures
      summary: 'Delete a lecture'
      description: 'Remove the specified lecture from storage'
      operationId: 6dbcc2ad9cd55a64f673d6f55ed826b0
      parameters:
        -
          name: id
          in: path
          description: 'ID of lecture to delete'
          required: true
          schema:
            type: integer
      responses:
        204:
          description: 'Lecture deleted'
        404:
          description: 'Lecture not found'
      security:
        -
          bearerAuth: []
...
  /api/topics:
    get:
      tags:
        - Topics
      summary: 'Get a list of topics'
      description: 'Gets a paginated list of topics with categories.'
      operationId: 8e16748db53ac966e8af74d772501056
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopicResource'
    post:
      tags:
        - Topics
      summary: 'Create a new topic'
      description: 'Creates a new topic with the provided data.'
      operationId: fead0ba561f17de0c4c6bccc6f8605e0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreTopicRequest'
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResource'
        409:
          description: 'Topic already exists'
      security:
        -
          bearerAuth: []
  '/api/topics/{id}':
    get:
      tags:
        - Topics
      summary: 'Get a specific topic'
      description: 'Gets details of a specific topic by ID.'
      operationId: 615c928c6685afe8fd57e2961667b24f
      parameters:
        -
          name: id
          in: path
          description: 'Topic ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResource'
        404:
          description: 'Topic not found'
    put:
      tags:
        - Topics
      summary: 'Update a specific topic'
      description: 'Updates a specific topic by ID with the provided data.'
      operationId: 008fde2a26f834347140fc61e91b92f2
      parameters:
        -
          name: id
          in: path
          description: 'Topic ID'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTopicRequest'
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResource'
        404:
          description: 'Topic not found'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Topics
      summary: 'Delete a specific topic'
      description: 'Deletes a specific topic by ID.'
      operationId: 37b43ac2b6d2aed36106fb9c32e65964
      parameters:
        -
          name: id
          in: path
          description: 'Topic ID'
          required: true
          schema:
            type: integer
      responses:
        204:
          description: 'Successful operation with no content'
        404:
          description: 'Topic not found'
      security:
        -
          bearerAuth: []
  '/api/topics/{topicId}/courses':
    get:
      tags:
        - Topics
      summary: 'Get Courses by Topic'
      description: 'Returns all courses under a specific topic'
      operationId: getCoursesByTopic
      parameters:
        -
          name: topicId
          in: path
          description: 'ID of the topic to return courses for'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseResource'
        404:
          description: 'Topic not found'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    StoreLectureRequest:
      title: 'Store Lecture Request'
      required:
        - chapter_id
        - title
        - video
      properties:
        chapter_id:
          description: 'ID of the associated chapter.'
          type: integer
          format: int64
        title:
          description: 'Title of the lecture.'
          type: string
        video:
          description: video.
          type: file
        thumbnail_img:
          description: 'Thumbnail image.'
          type: file
        resources:
          description: 'List of resources for the lecture.'
          type: array
          items:
            required:
              - title
              - link
            properties:
              title:
                description: 'Title of the resource.'
                type: string
              link:
                description: 'Link to the resource.'
                type: string
                format: uri
            type: object
      type: object
    StoreTopicRequest:
      title: 'Store Topic Request'
      description: 'Request body data for creating a new topic.'
      required:
        - category_id
        - name
      properties:
        category_id:
          description: 'ID of the category to which the topic belongs.'
          type: integer
        name:
          description: 'Name of the new topic.'
          type: string
          maxLength: 255
      type: object
...
    UpdateLectureRequest:
      title: 'Update Lecture Request'
      properties:
        title:
          description: 'Title of the lecture.'
          type: string
        video:
          description: 'Video file for the lecture in MP4, MOV, or AVI format.'
          type: string
          format: binary
        thumbnail_img:
          description: 'Thumbnail image for the lecture in JPEG, PNG, JPG, or GIF format.'
          type: string
          format: binary
        resources:
          description: 'List of updated resources for the lecture.'
          type: array
          items:
            required:
              - title
              - link
            properties:
              title:
                description: 'Title of the resource.'
                type: string
              link:
                description: 'Link to the resource.'
                type: string
                format: uri
            type: object
      type: object
...
    UpdateTopicRequest:
      title: 'Store Topic Request'
      description: 'Request body data for creating a new topic'
      required:
        - category_id
        - name
      properties:
        category_id:
          description: 'ID of the category to which the topic belongs.'
          type: integer
        name:
          description: 'Name of the new topic.'
          type: string
          maxLength: 255
      type: object
...
    CategoryResource:
      title: 'Category Resource'
      description: 'Represents a category resource.'
      properties:
        id:
          description: 'The unique identifier for the category.'
          type: integer
        parent_category:
          $ref: '#/components/schemas/CategoryResource'
        name:
          description: 'The name of the category.'
          type: string
        courses:
          description: 'List of courses associated with the category.'
          type: array
          items:
            $ref: '#/components/schemas/CourseResource'
        topics:
          description: 'List of topics associated with the category.'
          type: array
          items:
            $ref: '#/components/schemas/TopicResource'
      type: object
...
    CourseResource:
      title: 'Course Resource'
      description: 'Represents a course resource.'
      properties:
        id:
          description: 'The unique identifier for the course.'
          type: integer
        title:
          description: 'The title of the course.'
          type: string
        description:
          description: 'The description of the course.'
          type: string
        price:
          description: 'The price of the course.'
          type: number
        author:
          description: 'The author of the course.'
          type: string
        categories:
          description: 'List of categories associated with the course.'
          type: array
          items:
            $ref: '#/components/schemas/CategoryResource'
        reviews:
          description: 'List of reviews associated with the course.'
          type: array
          items:
            $ref: '#/components/schemas/CourseReviewResource'
      type: object
    CourseReviewResource:
      title: 'Course Review Resource'
      description: 'Represents a course review resource.'
      properties:
        id:
          description: 'The unique identifier for the review.'
          type: integer
        user:
          $ref: '#/components/schemas/UserResource'
        rating:
          description: 'The rating given in the review.'
          type: integer
        content:
          description: 'The content of the review.'
          type: string
        created_at:
          description: 'The timestamp when the review was created.'
          type: string
          format: date-time
      type: object
    LectureResource:
      title: 'Lecture Resource'
      description: 'Represents a lecture resource.'
      properties:
        id:
          description: 'The unique identifier for the lecture.'
          type: integer
        chapter_id:
          description: 'The unique identifier of the chapter associated with the lecture.'
          type: integer
        title:
          description: 'The title of the lecture.'
          type: string
        resources:
          description: 'A collection of resources related to the lecture.'
          type: array
          items:
            $ref: '#/components/schemas/LectureResourceResource'
        video:
          $ref: '#/components/schemas/LectureVideoResource'
        created_at:
          description: 'The date and time when the lecture was created.'
          type: string
          format: date-time
        updated_at:
          description: 'The date and time when the lecture was last updated.'
          type: string
          format: date-time
      type: object
    LectureResourceResource:
      title: LectureResourceModelResource
      description: 'Lecture resource model resource'
      properties:
        id:
          description: 'ID of the lecture resource'
          type: integer
        lecture_id:
          description: 'The ID of the lecture associated with this resource'
          type: integer
        title:
          description: 'Title of the lecture resource'
          type: string
        link:
          description: 'A link to the lecture resource'
          type: string
          format: uri
      type: object
      xml:
        name: LectureResourceModelResource
    LectureVideoResource:
      title: LectureVideoResource
      description: 'Lecture video resource'
      properties:
        data:
          properties:
            id:
              description: 'Unique identifier for the lecture video'
              type: integer
              example: 1
            url:
              description: 'URL of the video'
              type: string
              example: 'http://example.com/video.mp4'
            thumbnail_url:
              description: 'Thumbnail URL of the video'
              type: string
              example: 'http://example.com/thumbnail.jpg'
            duration:
              description: 'Duration of the video in seconds'
              type: integer
              example: 3600
            lecture_id:
              description: 'Foreign key to Lecture'
              type: integer
              example: 10
            created_at:
              description: 'Creation date of the video'
              type: string
              format: date-time
            updated_at:
              description: 'Last update date of the video'
              type: string
              format: date-time
          type: object
      type: object
      xml:
        name: LectureVideoResource
    TopicResource:
      title: 'Topic Resource'
      description: 'Represents a topic resource.'
      properties:
        id:
          description: 'The unique identifier for the topic.'
          type: integer
        name:
          description: 'The name of the topic.'
          type: string
        categories:
          description: 'List of categories associated with the topic.'
          type: array
          items:
            $ref: '#/components/schemas/CategoryResource'
      type: object
    UserResource:
      title: 'User Resource'
      description: 'Represents a user resource.'
      properties:
        id:
          description: 'The unique identifier for the user.'
          type: integer
        display_name:
          description: 'The display name of the user.'
          type: string
        email:
          description: 'The email address of the user.'
          type: string
          format: email
        created_at:
          description: 'The timestamp when the user was created.'
          type: string
          format: date-time
        updated_at:
          description: 'The timestamp when the user was last updated.'
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Login with email and password from user module to get the authentication token'
      name: Authorization
      in: header
      scheme: bearer
...